Pull out sso-common/ sources
