+++ sso-common/README.md
@@ -1,0 +1,4 @@
+### Common source for "sso" (sunrise-sunset.org) implementations
+
+**Heads up:** This is not a module, this is just a directory for source used by both the `:sso` and `:sso-backport`
+modules. The common source is agnostic to any date/time types.
