+++ sso-common/src/main/java/drewhamilton/skylight/sso/network/SsoDate.kt
@@ -1,8 +1,5 @@
 package drewhamilton.skylight.sso.network
 
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
-
 /**
  * Indicates a String in the sunrise-sunset.org API's date format, e.g. "2019-05-21"
  */
@@ -16,10 +13,3 @@ import java.time.format.DateTimeFormatter
     AnnotationTarget.VALUE_PARAMETER
 )
 annotation class SsoDate
-
-@Suppress("NewApi")
-@SsoDate
-internal fun LocalDate.toSsoDateString() = DATE_FORMATTER.format(this)
-
-@Suppress("NewApi")
-private val DATE_FORMATTER: DateTimeFormatter by lazy { DateTimeFormatter.ISO_LOCAL_DATE }
