+++ sso-common/src/main/java/drewhamilton/skylight/sso/network/response/Response.kt
