+++ sso-common/src/main/java/drewhamilton/skylight/sso/network/response/SunriseSunsetInfo.kt
